Adding custom JSON marshal response
