+++ go/http_test.go
@@ -2,16 +2,33 @@ package common_test
 
 import (
 	"net/http"
+	"strings"
 	"testing"
 
 	common "github.com/Cyb3r-Jak3/common/go"
 )
+
+type requestBody struct {
+	Name string `json:"name"`
+}
+
+type responseBody struct {
+	Test requestBody `json:"url"`
+}
 
 // Hello is a simple hello function
 func StringTest(w http.ResponseWriter, _ *http.Request) { common.StringResponse(w, "Hello") }
 
 func JSONTest(w http.ResponseWriter, _ *http.Request) {
 	common.JSONResponse(w, []byte(`"hello": "world"`))
+}
+
+func JSONMarshalTest(w http.ResponseWriter, _ *http.Request) {
+	common.JSONMarshalResponse(w, &requestBody{Name: "test"})
+}
+
+func JSONMarshalBadTest(w http.ResponseWriter, _ *http.Request) {
+	common.JSONMarshalResponse(w, nil)
 }
 
 func ContentTest(w http.ResponseWriter, _ *http.Request) {
@@ -33,8 +50,72 @@ func TestJsonResponse(t *testing.T) {
 	checkResponse(t, rr, http.StatusOK)
 }
 
+func TestJSONMarshall(t *testing.T) {
+	r, err := http.NewRequest("GET", "/", nil)
+	if err != nil {
+		t.Errorf("Wanted no error and got %s", err)
+	}
+	rr := executeRequest(r, JSONMarshalTest)
+	resp := rr.Result()
+	if resp.Header.Get("Content-Type") != common.JSONApplicationType {
+		t.Errorf("Wanted JSON response and got %s", resp.Header.Get("Content-Type"))
+	}
+	checkResponse(t, rr, http.StatusOK)
+}
+
 func TestContentResponse(t *testing.T) {
 	r, _ := http.NewRequest("GET", "/", nil)
 	rr := executeRequest(r, ContentTest)
 	checkResponse(t, rr, http.StatusOK)
+}
+
+func TestEmptyResponse(t *testing.T) {
+	resp, err := common.DoJSONRequest(
+		"", "https://httpbin.org/anything", &requestBody{Name: "value"}, nil,
+	)
+	if err != nil {
+		t.Errorf("Wanted no error and got %s", err)
+	}
+	if resp.StatusCode != 200 {
+		t.Errorf("Wanted status code and got %d", resp.StatusCode)
+	}
+}
+
+func TestEmptyBody(t *testing.T) {
+	resp, err := common.DoJSONRequest(
+		"", "https://httpbin.org/anything", nil, nil,
+	)
+	if err != nil {
+		t.Errorf("Wanted no error and got %s", err)
+	}
+	if resp.StatusCode != 200 {
+		t.Errorf("Wanted status code and got %d", resp.StatusCode)
+	}
+}
+
+func TestErrors(t *testing.T) {
+	resp, err := common.DoJSONRequest(
+		"", "example.com", nil, nil,
+	)
+	if err == nil {
+		t.Error("Got no error and wanted one")
+	} else if !strings.Contains(err.Error(), "unsupported protocol scheme") {
+		t.Errorf("Wanted bad protocol scheme and got %s", err)
+	}
+	if resp != nil {
+		t.Error("Wanted empty response and it was not")
+	}
+}
+
+func TestGoodRequest(t *testing.T) {
+	var respBody responseBody
+	resp, err := common.DoJSONRequest(
+		"GET", "https://httpbin.org/anything", &requestBody{Name: "Hello"}, &respBody,
+	)
+	if err != nil {
+		t.Errorf("Wanted no error and got %s", err)
+	}
+	if resp.StatusCode != 200 {
+		t.Errorf("Wanted status code and got %d", resp.StatusCode)
+	}
 }
